chore: Split `closeHttpClient` to `src-modern` and `src-legacy` with different implementations
