+++ requests/src-modern/src-modern/CompatUtil.scala
@@ -1,0 +1,10 @@
+package requests
+
+import java.net.http.HttpClient
+
+private object CompatUtil {
+  // For JDK 21+ and Scala Native
+  @inline
+  final def closeHttpClient(httpClient: HttpClient): Unit =
+    httpClient.close()
+}
