+++ requests/src/requests/Requester.scala
@@ -107,47 +107,11 @@ object BaseSession {
    * Closes an HttpClient, using reflection to handle both Java 21+ (which has close()) and earlier
    * versions (which require accessing internal selector).
    */
-  def closeHttpClient(httpClient: HttpClient): Unit = {
-    try {
-      val closeMethod = classOf[HttpClient].getMethod("close")
-      closeMethod.invoke(httpClient)
-    } catch {
-      case _: NoSuchMethodException =>
-        // Java < 21: use reflection to access internal selectorManager and close its selector
-        // HttpClient.newBuilder().build() returns HttpClientFacade which wraps HttpClientImpl
-        try {
-          val facadeClass = httpClient.getClass
-          val implField = facadeClass.getDeclaredField("impl")
-          implField.setAccessible(true)
-          val impl = implField.get(httpClient)
-          val selectorManagerField = impl.getClass.getDeclaredField("selmgr")
-          selectorManagerField.setAccessible(true)
-          val selectorManager = selectorManagerField.get(impl)
-          // SelectorManager has a 'selector' field we can close
-          val selectorField = selectorManager.getClass.getDeclaredField("selector")
-          selectorField.setAccessible(true)
-          val selector = selectorField.get(selectorManager)
-          val closeMethod = selector.getClass.getMethod("close")
-          closeMethod.invoke(selector)
-        } catch {
-          case _: Exception =>
-            System.err.println(
-              "requests: Unable to close HttpClient SelectorManager thread. " +
-              "To fix thread leaks on Java <21, add JVM arg: " +
-              "--add-opens java.net.http/jdk.internal.net.http=ALL-UNNAMED",
-            )
-        }
-    }
-  }
+  def closeHttpClient(httpClient: HttpClient): Unit = CompatUtil.closeHttpClient(httpClient)
 }
 
 object Requester {
   val officialHttpMethods = Set("GET", "POST", "HEAD", "OPTIONS", "PUT", "DELETE", "TRACE")
-  private lazy val methodField: java.lang.reflect.Field = {
-    val m = classOf[HttpURLConnection].getDeclaredField("method")
-    m.setAccessible(true)
-    m
-  }
 }
 
 case class Requester(verb: String, sess: BaseSession) {
